persist: exempt strings that look like ISO8601 timestamps from trimming
This fixes a bug that Chuck and Paul were hitting where a string
timestamp column with a lower of "2023-08-19T12:00:00.000Z" and an upper
of "2023-08-20T12:00:00.000Z" was getting truncated to "2023-08-" and
"2023-08." by trimming.
