+++ src/persist-types/Cargo.toml
@@ -12,6 +12,7 @@ publish = false
 anyhow = { version = "1.0.66", features = ["backtrace"] }
 arrow2 = { version = "0.16.0", features = ["compute_aggregate", "io_ipc", "io_parquet"] }
 bytes = "1.3.0"
+chrono = { version = "0.4.23", default-features = false, features = ["std"] }
 hex = "0.4.3"
 mz-ore = { path = "../ore", features = ["test"] }
 mz-proto = { path = "../proto" }
